# Enable Precompiled Headers for faster compilation