--- a/base/cinnamon.vcxproj.filters
+++ b/base/cinnamon.vcxproj.filters
@@ -545,5 +545,8 @@
     <ClCompile Include="..\dependencies\minhook\hde\hde64.c">
       <Filter>dependencies\minhook\hde32</Filter>
     </ClCompile>
+    <ClCompile Include="pch.cpp">
+      <Filter>source</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>