--- a/base/pch.cpp
+++ b/base/pch.cpp
@@ -1,0 +1,1 @@
+#include "pch.h"