--- a/base/pch.h
+++ b/base/pch.h
@@ -1,0 +1,1 @@
+#pragma once