--- a/base/pch.h
+++ b/base/pch.h
@@ -1,1 +1,16 @@
 #pragma once
+
+#ifdef _WIN32
+#include <Windows.h>
+#endif
+
+#ifdef __cplusplus
+#include <filesystem>
+#include <fstream>
+
+#ifdef _WIN32
+#include <directxmath.h>
+#endif
+
+#include "common.h"
+#endif