--- a/base/pchc.c
+++ b/base/pchc.c
@@ -1,0 +1,1 @@
+#include "pch.h"