This PR improves compilation times. Results of a clean build on a 4-core CPU:
- **Release**:
  - before: **56531 ms**
  - after: **32473 ms**
- **Debug**:
  - before: **39149 ms**
  - after: **13188 ms**