@rollraw 2021-06-25
hey Daniel, good job.
one thing i care about is visual studio bug when cheat could be compiled wrong and only after full recompilation will be fine. it's caused by cached obj files, so did you expirienced smth similar with precached headers?
btw what you used to benchmark it?