@danielkrupinski 2021-06-25
Generally you precompile headers that are unlikely to change often (e.g. standard library, Windows headers) and also PCH files are different from OBJ files, so I think that won't be the case here.
For measuring build times: https://stackoverflow.com/questions/82128/displaying-build-times-in-visual-studio
For the analysis I used **vcperf** and **Windows Performance Analyzer** - here is the guide https://docs.microsoft.com/en-us/cpp/build-insights/tutorials/vcperf-and-wpa?view=msvc-160